requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.9.1.post1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
fake-useragent==1.1.1
flake8==6.0.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
importlib-resources==5.10.1
lxml==4.9.2
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.3
ndg-httpsclient==0.5.1
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyOpenSSL==22.1.0
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.4.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
telegraph==2.2.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
yarl==1.8.2
zipp==3.11.0
